Fixed some compile errors

+++ dtbtool.c
@@ -530,11 +530,11 @@ int main(int argc, char **argv)
     close(out_fd);
 
     if (expected != wrote) {
-        log_err("error writing output file, please rerun: size mismatch %d vs %d\n",
-                expected, wrote);
+        log_err("error writing output file, please rerun: size mismatch %u vs %u\n",
+                (uint)expected, (uint)wrote);
         rc = RC_ERROR;
     } else
-        log_dbg("Total wrote %u bytes\n", wrote);
+        log_dbg("Total wrote %u bytes\n", (uint)wrote);
 
     if (rc != RC_SUCCESS)
         unlink(output_file);
